tests/wpt/web-platform-tests/sanitizer-api/sanitizer-sanitizeToString.https.tentative.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/testcases.sub.js"></script>
</head>

<body>
<script>
    testcases.forEach(c => test(t => {
        let s = new Sanitizer(c.config_input);
        assert_equals(s.sanitizeToString(c.value), c.result);
    }, "SanitizerAPI sanitizeToString function for " + c.message));
</script>
</body>
</html>
